Reorganized structure of files.
